+++ src/implementations/checkdigit/historical.scopes.ts
@@ -1,7 +1,5 @@
 // scopes that were registered in old Alpha; all have check digits and hash length of 4.
 export enum AlphaHashidScopes {
-  Buyer = 'Buyer',
-  Supplier = 'Supplier',
   PurchaseOrder = 'PurchaseOrder',
   Shipment = 'Shipment',
   User = 'User'